refactor: file strucutre
