+++ packages/app/e2e/settings/settings-language.spec.ts
@@ -1,5 +1,5 @@
-import { test, expect } from "./fixtures"
-import { modKey, settingsLanguageSelectSelector } from "./utils"
+import { test, expect } from "../fixtures"
+import { modKey, settingsLanguageSelectSelector } from "../utils"
 
 test("smoke changing language updates settings labels", async ({ page, gotoSession }) => {
   await page.addInitScript(() => {
@@ -25,7 +25,6 @@ test("smoke changing language updates settings labels", async ({ page, gotoSessi
   const heading = dialog.getByRole("heading", { level: 2 })
   await expect(heading).toHaveText("General")
 
-
   const select = dialog.locator(settingsLanguageSelectSelector)
   await expect(select).toBeVisible()
   await select.locator('[data-slot="select-select-trigger"]').click()
@@ -34,11 +33,7 @@ test("smoke changing language updates settings labels", async ({ page, gotoSessi
 
   await expect(heading).toHaveText("Allgemein")
 
-
   await select.locator('[data-slot="select-select-trigger"]').click()
   await page.locator('[data-slot="select-select-item"]').filter({ hasText: "English" }).click()
   await expect(heading).toHaveText("General")
-
-
-
 })
